+++ tilings/strategies/verification.py
@@ -14,7 +14,7 @@ from comb_spec_searcher import (
 )
 from comb_spec_searcher.exception import InvalidOperationError, StrategyDoesNotApply
 from comb_spec_searcher.typing import Objects, Terms
-from permuta import Perm
+from permuta import Av, Perm
 from permuta.permutils import (
     is_insertion_encodable_maximum,
     is_insertion_encodable_rightmost,
@@ -177,11 +177,17 @@ class OneByOneVerificationStrategy(BasisAwareVerificationStrategy):
         )
 
     def verified(self, comb_class: Tiling) -> bool:
-        return comb_class.dimensions == (1, 1) and (
-            frozenset(ob.patt for ob in comb_class.obstructions) not in self.symmetries
-            or any(
-                isinstance(ass, ComponentAssumption) for ass in comb_class.assumptions
-            )
+        if not comb_class.dimensions == (1, 1):
+            return False
+        if not self.basis:
+            return True
+        tiling_class = Av([ob.patt for ob in comb_class.obstructions])
+        sym_classes = (Av(sym) for sym in self.symmetries)
+        is_strict_subclass = any(
+            tiling_class.is_subclass(cls) and cls != tiling_class for cls in sym_classes
+        )
+        return is_strict_subclass or any(
+            isinstance(ass, ComponentAssumption) for ass in comb_class.assumptions
         )
 
     def get_genf(
@@ -225,7 +231,9 @@ class OneByOneVerificationStrategy(BasisAwareVerificationStrategy):
         )
 
     def __str__(self) -> str:
-        return "one by one verification"
+        if not self.basis:
+            return "one by one verification"
+        return f"One by one subclass of {Av(self.basis)}"
 
 
 class DatabaseVerificationStrategy(TileScopeVerificationStrategy):
@@ -442,7 +450,7 @@ class LocalVerificationStrategy(TileScopeVerificationStrategy):
     localized, i.e. in a single cell and the tiling is not 1x1.
     """
 
-    def __init__(self, ignore_parent: bool = True, no_factors: bool = False):
+    def __init__(self, ignore_parent: bool = False, no_factors: bool = False):
         self.no_factors = no_factors
         super().__init__(ignore_parent=ignore_parent)
 
@@ -533,7 +541,7 @@ class InsertionEncodingVerificationStrategy(TileScopeVerificationStrategy):
     Verify all n x 1 and 1 x n tilings that have a regular insertion encoding.
     """
 
-    def __init__(self, ignore_parent: bool = True):
+    def __init__(self, ignore_parent: bool = False):
         super().__init__(ignore_parent=ignore_parent)
 
     def pack(self, comb_class: Tiling) -> StrategyPack:
@@ -608,7 +616,7 @@ class MonotoneTreeVerificationStrategy(TileScopeVerificationStrategy):
     Verify all tiling that is a monotone tree.
     """
 
-    def __init__(self, ignore_parent: bool = True, no_factors: bool = True):
+    def __init__(self, ignore_parent: bool = False, no_factors: bool = True):
         self.no_factors = no_factors
         super().__init__(ignore_parent=ignore_parent)