+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- One by one verification will now only verify subclasses of the given basis.
+- Verification strategies no longer ignore parent
+
 ## [3.0.0] - 2021-06-14
 ### Added
 - Rearrange assumption strategy